feat(checkbox): Allow children and custom margin
We want to be able to put more complex elements in the label than
just text. We also want to be able to specify the margin rather than use
the rather luxurious default spacing.

This also fixes a missing flex-shrink setting that would make skinny checkboxes
at narrow screen widths.