fix(SearchInput): Forward ref
When we switched to styled components, it appears we lost innerRef as a tool to pass a function down
to a styled element that would be called in order to set a ref to the element. This broke Algolia
search, which was using the functionality to blur the search input on submit. The React docs
recommend using React.forwardRef() to allow forwarding a ref passed down from a parent to a child,
so that's what we do. Styled components uses the forwarded ref as a normal ref, making the input
element available to the parent.